Revert "run graalvm native tests github action"
This reverts commit f73f40f21a45db996a393057f957e3c816f5fef8.
