+++ pom.xml
@@ -277,8 +277,6 @@
                 <include>**/RSMetaDataTest</include>
                 <!--  SegFault  -->
                 <exclude>**/UDFCustomErrorTest.java</exclude>
-                <!--  ${project.name} not properly replaced on Windows and MacOS in CI  -->
-                <exclude>**/DBMetaDataVersionTest.java</exclude>
               </excludes>
             </configuration>
           </plugin>
